requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
transformers>=4.36.0
shap>=0.43.0
fhirclient>=3.2.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
aiohttp>=3.9.0
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0
flake8>=6.1.0 